Đăng ký bác sĩ

+++ src/main/java/com/do_an/appointment/dtos/DoctorDTO.java
@@ -16,5 +16,6 @@ public class DoctorDTO {
     private String specialty;
 
     @JsonProperty("experience")
-    private int experience;
+    private Long experience;
+
 }
